requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
eventlet==0.35.2
psycopg2-binary==2.9.9
python-docx==1.1.0
spacy==3.7.2
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
transformers>=4.46.0
accelerate>=0.26.0
bitsandbytes>=0.41.0
huggingface_hub>=0.20.0
redis>=5.0.0
protobuf
